test: cql-pytest: give it a valid Python package name
Using regular import mechanisms and running static code analysis tools
is problematic because `cql-pytest` is not a valid Python package name.
Currently the codebase resorts to `sys.path` modification hacks for
importing.

Fix the name by replacing the dash with an underscore.

Update all references in the codebase (including comments and docs) to
use the new name.