test: scylla-gdb: give it a valid Python package name
As in the previous commit, replace the dash with an underscore to make
it possible to use regular Python import mechanisms and static code
analysis tools.

The tool itself is still named `scylla-gdb.py`, it probably deserves to
be renamed eventually as well, but that's out of the scope of this
patchset.