test: run{-cassandra}.py: get the script's dir using a more general method
The run scripts were using `sys.path[0]` to obtain the directory
containing the scripts in order to pass them to `pytest` (because the
directories contain tests for `pytest` to run).

Use a different method: `pathlib.Path(__file__).parent.resolve()`. This
method will work even if we run the script as a Python submodule, e.g.
`python -m test.cql_pytest.run` from repository top level dir (in this
case `sys.path[0]` will point to the repository top level).

Note: the script is not a valid submodule of the `test` package yet
because of broken imports, but this commit prepares us for later commits
which will change this.