utils/logalloc: introduce segment_store_backend
We want to make it possible to switch between segment store
implementations run-time but without making all members of the segment
store virtual. The role of the segment stores is to allocate/deallocate
and assign ids to segments. The seastar allocator based one does this
based on a memory-layout description obtained from the seastar
allocator, while the standard allocator based one maintains a vector of
segments to provide the same services. We observe that it is possible to
standardize on the former if we also allocate a large contiguous memory
area for segments when using the standard allocator. We achieve that by
simply allocating a large aligned buffer and abstract the way segments
are allocated/deallocated with a virtualized segment store backend.
The plan is that we will reuse the accounting code of the seastar
allocator based segment store largely untouched, with a virtual call
needed (to the store backend) only when actually allocating/deallocating
segments.
This patch only introduces the segment store backend interface and
implementations, they are not used yet.