wasm: reuse instances across groups
If a UDF is called in a scheduling group and no instance has been
cached for that group, but there is an instance of this UDF for
another group that is not currently used, the instance is now
taken over by the group where the UDF is currently being used.
This lets us avoid a situation where an instance is created
fo a UDF while another instance for the same UDF gets evicted
due to a timeout. If the instance weren't going to be evicted
anyway, the amortized cost of calling the UDF in these two
scheduling gropus is the same as if the instance was not taken
over - one instance is reused and one new instance is created.

A test that confirms the reuse is added to boost/user_function_test