raft server, log size limit in bytes
Before this patch we could get an OOM if we
received several big commands. The number of
commands was small, but their total size
in bytes was large.

max_snapshot_trailing_bytes is needed to guarantee
progress. Without this limit the fsm could
get stuck if the size of the next item is greater than  max_log_size_in_bytes - snapshot_trailing_bytes.