sstables: index_entry: add use_caching::cache_locally
The semantics of BYPASS CACHE are too strict for the goal of avoiding cache
pollution with long-term index. BYPASS CACHE caches index pages for the lifetime
of the reader, but does not cache index file pages at all. This can potentially
cause additional I/O, as described in the commit introducing
global_cache_index_pages. The middle ground is to also cache index file pages
for the lifetime of the reader. This patch makes it happen.