cql3: restrictions: add analyzed_where_clause
analyzed_where_clause is a class whose job is
to analyze the WHERE clause and generate
the query plan. It decides whether to use
indexing, filtering, generates partition
and clustering ranges, etc.

In the future it will replace statement_restrictions.

It has two fields.
One keeps the old statement_restrictions class.
The other one keeps the new representation.

All functionality will be gradually implemented
using the new representation, but for now
all analyzed_where_clause methods call the corresponding
methods in statement_restrictions.
Methods will be implemented one by one,
which allows to divide the transition into small
testable and reviewable chunks.

The new representation is a variant:
Either a query that fetches data from a single table,
or a query that uses an index.

I'm hoping that we could express the index query
as two separate queries to the respective tables.
This would drastically reduce complexity,
but in this PR I only focus on single table
partition key restrictions to limit patch size.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>