gms: feature_service: recalculate supported features on config changes and notify
We add a mechanism to `feature_service` which recalculates the set of
supported features (actually, the internally stored set is the set of of
'disabled' features - supported features are the complement of disabled
ones) and calls registered notification callbacks when the database
configuration changes.

To do that, we observe the `experimental_features` field of
`db::config&`; this field is the one that affects the set of features
most significantly. This is doesn't call all potential changes as I
explained in a comment, but it will cover the SUPPORTS_RAFT feature
which this feature is intended for.

The mechanism allows the node to start supporting features in runtime,
but only the ones which were explicitly listed in a
`can_be_supported_dynamically` set. For now I left the set empty; I
tested manually that it does work for enabling features but no features
are actually prepared yet to be enabled like that. SUPPORTS_RAFT is
close, but some small changes are still required in the group 0 code
(right now the feature listener for SUPPORTS_RAFT is only installed if
the Raft experimental feature was enabled during startup). I plan to
send them in a follow-up PR.