restrictions: add partition_key_restrictions
This commit adds a class responsible for handling
partition key restrictions.

A while ago we got rid of partition_key_restrictions
and clustering_key_restrictions classes, but the
reason for that was that they were a part of a
convoluted and unelastic inheritance hierarchy.

The idea of splitting responsibility
for handling partition and clustering restrictions
into two modules itself was good.

Extracting partition and clustering ranges
have they own rules, quirks and complexities
and it's useful to have this code in a separate
place.

Over time as we relax the grammar more and more
the code there will grow, but the basic
functionality remains the same - these classes
take an expression, generate ranges and
decide what expressions to use for filtering.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>